# backend/openApi/reference/mapAdmin.v1.yaml
openapi: 3.0.0
servers:
  - description: mapadmin
    url: 'http://localhost:3000/api'
info:
  description: |-
    Все запросы кроме getAuthData, возвращают стандартный ответ вида
    {
      "isError":boolean,
      "response": string or object or array
    },
    В случае ошибки isError будет true, в response будет описание ошибки. Соответственно, в случае успеха isError будет false.
    Если пытаемся получить доступ к конечной точке, для другой роли. Получим отывет
    {
      "isError": true,
      "response": "No permission"
    }
    Если пытаемся получить доступ к конечной точке, без авторизации. Получим отывет
    {
      "isError": true,
      "response": "Unauthorized"
    }
  version: 1.0.0
  title: mapadmin API
  contact: {}
  license: {}
...
paths:
  /login:
    post:
      tags:
        - Авторизация
      summary: Запрос авторизации
      operationId: login
      requestBody:
        description: |
          Отправляем данные для авторизации
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - login
                - password
              properties:
                login:
                  type: string
                  format: email
                  maxLength: 320
                password:
                  type: string
                  minimum: 8
                  maximum: 72
      responses:
        '200':
          description: Ответ
          content:
            application/json:
              schema:
                type: object
                properties:
                  isError:
                    type: boolean
                  response:
                    type: string
                required:
                  - isError
                  - response
              examples:
                Успешный вход:
                  value:
                    isError: false
                    response: OK
                Неверный логин или пароль:
                  value:
                    isError: true
                    response: Unauthorized
        '400':
          description: Если какой то параметр не будет отправлен
          content:
            application/json:
              schema:
                type: string
                example: Bad Request
      description: Вход в аккаунт
...
  /admin/addUser:
    post:
      tags:
        - admin
      summary: Добавление пользователей
      operationId: addUser
      description: |
        Добавление пользователей с различными правами
      requestBody:
        description: |
          Отправляем логин, пароль и id прав
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  maximum: 320
                password:
                  type: string
                  minimum: 8
                  maximum: 72
                permission_id:
                  type: integer
              required:
                - email
                - password
                - permission_id
      responses:
        '200':
          description: Возвращает данные добавленного пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  isError:
                    type: boolean
                  response:
                    oneOf:
                      - type: string
                      - $ref: ../models/getUsers.v1.json
                required:
                  - isError
                  - response
              examples:
                Пользователь добавлен:
                  description: response возвращает поля добавленного пользователя
                  value:
                    isError: false
                    response:
                      - id: 1
                        email: user@user.com
                        permission: user
                Такой пользователь уже есть:
                  value:
                    isError: true
                    response: this user already exists
                поле email пустое или не строкового типа:
                  value:
                    isError: true
                    response: email must not be empty
                Не авторизован:
                  value:
                    isError: true
                    response: Unauthorized
                Поле пароль не отправленно:
                  value:
                    isError: true
                    response: 'ValidationError: password: is a required property'
                Поле пароль слишком короткое:
                  value:
                    isError: true
                    response: 'ValidationError: password: should NOT be shorter than 8 characters'
                Неверный формат email:
                  value:
                    isError: true
                    response: 'ValidationError: email: should match format "email"'
                Данного permission_id нет в базе:
                  value:
                    isError: true
                    response: permission with this id not found
                'permission_id пустое, либо не является числом':
                  value:
                    isError: true
                    response: permission_id must be integer
  /admin/editUser:
    post:
      tags:
        - admin
      summary: Изменение логина и пароля
      operationId: editUser
      description: "Изменение логина(email) и пароля, если пароль пустая строка или пароль не отправлен, аналогично и с логином. \n"
      requestBody:
        description: |
          Отправляем логин и пароль
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - id
              properties:
                id:
                  type: integer
                email:
                  type: string
                  format: email
                  maximum: 320
                password:
                  type: string
                  minimum: 8
                  maximum: 72
      responses:
        '200':
          description: Возвращает данные отредактированного пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  isError:
                    type: boolean
                  response:
                    oneOf:
                      - type: string
                      - $ref: ../models/getUsers.v1.json
              examples:
                Пароль изменен:
                  description: response возвращает поля отредактированного пользователя
                  value:
                    isError: false
                    response:
                      - id: 1
                        email: user@user.com
                        permission: user
                Запрос не из под администратора:
                  value:
                    isError: true
                    response: No permission
                Не авторизован:
                  value:
                    isError: true
                    response: Unauthorized
                Поле пароль не отправленно:
                  value:
                    isError: true
                    response: 'ValidationError: password: is a required property'
                Поле пароль слишком короткое:
                  value:
                    isError: true
                    response: 'ValidationError: password: should NOT be shorter than 8 characters'
                Неверный формат email:
                  value:
                    isError: true
                    response: 'ValidationError: email: should match format "email"'
                Остальные ошибки:
                  value:
                    isError: true
                    response: fail
                Поле id неверного формата или отсутствует:
                  value:
                    isError: true
                    response: userId must be not empty
...
  /admin/loginAs:
    post:
      summary: Зайти под другим пользователем
      tags:
        - admin
      responses:
        '200':
          description: OK
          headers: {}
          content:
            application/json:
              schema:
                type: object
                properties:
                  isError:
                    type: boolean
                  response:
                    type: string
              examples:
                Пример:
                  value:
                    isError: false
                    response: OK
        '':
          content:
            application/json:
              schema:
                type: object
                properties:
                  isError:
                    type: boolean
                  response:
                    type: string
              examples:
                Успех:
                  value:
                    isError: false
                    response: OK
                Нет такого id:
                  value:
                    isError: true
                    response: this userId not found
          description: Ответ
      operationId: get-admin-loginAs
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type:
                    - string
                    - integer
            examples: {}
        description: Отправлем id пользователя под которого мы хотим зайти
      description: 'Заходим под другим пользователем, система будет идентифицировать как другого пользователя до того как не будет выполнен запрос returnToAdmin'
...
  /user/addPoint:
    post:
      summary: Добавление точки в базу данных
      operationId: post-user-addPoint
      responses:
        '200':
          description: 'При успехе вернет данные добавленной точки, при дубликате, данные дубликатов.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  isError:
                    type: boolean
                  response:
                    oneOf:
                      - type: string
                      - $ref: ../models/getPointsUser.v1.json
                      - $ref: ../models/duplicate.v1.json
                required:
                  - isError
                  - response
              examples:
                Успешно:
                  value:
                    isError: false
                    response:
                      - id: 1223
                        full_city_name: 'Майкоп, Республика Адыгея, Россия'
                        street: улица Карла Маркса
                        house: '23'
                        title: Зоомир
                        lng: 44.6117
                        lat: 40.0888
                        apartment: '77'
                        hours: 'Пн-Пт: 9.00-20.00; Сб-Вс: 8.00-19.00'
                        phone: '656565'
                        site: 'https://karmy.su'
                        isActive: 1
                        description: ''
                        timeStamp: '2020-09-04 17:41:24'
                        moder_status: accept
                Дубликат:
                  value:
                    isError: true
                    response:
                      duplicate:
                        points:
                          - full_city_name: 'Майкоп, Республика Адыгея, Россия'
                            street: Заводская улица
                            house: '55'
                            title: магазин
                            lng: 44.598073
                            lat: 40.123267
                            apartment: null
                            hours: sadasdasdasdasdas
                            phone: '24323423434'
                            site: qwewqewqe
                            user: user@user.user
                          - full_city_name: 'Майкоп, Республика Адыгея, Россия'
                            street: Заводская улица
                            house: '55'
                            title: магазин
                            lng: 44.598073
                            lat: 40.123267
                            apartment: null
                            hours: sadasdasdasdasdas
                            phone: '24323423434'
                            site: qwewqewqe
                            user: user@user.user
                        point:
                          title: магазин
                          hours: sadasdasdasdasdas
                          phone: '24323423434'
                          site: qwewqewqe
                          lat: 40.123267
                          lng: 44.598073
                          full_city_name: 'Майкоп, Республика Адыгея, Россия'
                          city: Майкоп
                          street: Заводская улица
                          house: '55'
                Пустые координаты:
                  value:
                    isError: true
                    response: lat and lng must not be empty
                Проблема с получением гео данных(скорее всего в данной точке нет строений):
                  value:
                    isError: true
                    response: failed to get geodata
      tags:
        - user
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                lng:
                  type: number
                lat:
                  type: number
                title:
                  type: string
                apartment:
                  type: string
                hours:
                  type: string
                phone:
                  type: string
                site:
                  type: string
                description:
                  type: string
                force:
                  type: boolean
                  description: 'при true добавит точку в любом случае, даже если есть дубликат'
                isActive:
                  type: boolean
              required:
                - lng
                - lat
      description: 'Добавляет точку в базу данных, статус выставляется moder_status выставляется в moderated, для дальнейшего одобрения модератором. Пока статус не выставится в accept точке не будет отображаться на карте. При добавлении точки с похожими координаторами (точка не дальше 200 метров от имеющейся), будет возвращена ошибка. Для того чтоб все равно добавить точку нужно выставить поле force в запросе в true. Все точки подходящие под критерий в базе будут помечены как возможные дубликаты и присвоены к одной группе. В дальнейшем если точка будет близкой хотя бы к одной из группы, как дубликаты будут выданы все точки группы.'
...
  /user/delPoint:
    post:
      summary: Удаление точки
      tags:
        - user
      responses:
        '200':
          description: Успех
          content:
            application/json:
              schema:
                type: object
                properties:
                  isError:
                    type: boolean
                  response:
                    type: string
                required:
                  - isError
                  - response
              examples:
                Успех:
                  value:
                    isError: false
                    response: OK
                Нет такой точки:
                  value:
                    isError: true
                    response: point id not found
                Поле id некорректно:
                  value:
                    isError: true
                    response: pointId must not be empty
      operationId: get-user-delPoint
      description: 'Удаляет точку принадлежащию пользователю, который авторизован в данный момент. Принимает в качестве параметра id точки В случае если данный id, принадлежит точке другого пользователя или не существует будет возвращена ошибка с текстом "point id not found" Если в поле отосланы некорректные данные или поле не отправлено будет возвращена ошибка с текстом "pointId must not be empty"'
      parameters: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
              required:
                - id
        description: Отправляем id точки
...
  /moder/setPointRefuse:
    post:
      summary: Отклоняет точку
      tags:
        - moder
      responses:
        '200':
          description: Ответ
          content:
            application/json:
              schema:
                type: object
                properties:
                  isError:
                    type: boolean
                  response:
                    type:
                      - string
                      - integer
                    description: id точки в случае успеха
                required:
                  - isError
                  - response
              examples:
                Успех:
                  value:
                    isError: false
                    response: 33
                Некоректное поле id:
                  value:
                    isError: true
                    response: pointId must not be empty
                отправлен id несуществующей точки:
                  value:
                    isError: true
                    response: fail
      operationId: get-moder-setPointRefuse
      description: 'Модератор помечает точку как отклоненную(refuse), он так же может оставить комментарий, объяснив причину отказа. '
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  description: id точки
                description:
                  type: string
                  description: комментарий к отказу
              required:
                - id
  /moder/setPointAccept:
    post:
      summary: Подтверждает точку
      tags:
        - moder
      responses:
        '200':
          description: Ответ
          content:
            application/json:
              schema:
                type: object
                properties:
                  isError:
                    type: boolean
                  response:
                    type:
                      - string
                      - integer
                    description: id точки в случае успеха
                required:
                  - isError
                  - response
              examples:
                Успех:
                  value:
                    isError: false
                    response: 33
                Некоректное поле id:
                  value:
                    isError: true
                    response: pointId must not be empty
                отправлен id несуществующей точки:
                  value:
                    isError: true
                    response: fail
      operationId: get-moder-setPointAccept
      description: 'Модератор помечает точку как подтвержденную(accept). '
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  description: id точки
              required:
                - id
        description: ''
...